Comb creation only first harmonic
This code is a complete pipeline for generating combs given any input IR spectra. Only the first harmonic is returned. The output csv file is written in blocks of two rows at a time. In each pair, the first row is frequencies, and the second row is the comb height at that given point.